Simplify Material theme
Remove sizing and spacing related custom properties and calculations

Align typography naming to Material specs. Remove responsive typography
and font size calculations.

Flatten the folder structure so that the primarily used, element
specific, imports are at the top level. Remove the main `material.html`
import as it served little purpose. Consider reintroducing it as a
bundle import in the future, if wanted.

Small adjustments to date picker styles (depends on
https://github.com/vaadin/vaadin-date-picker/pull/393)